refactor: delete unused CreateProfileScreenModel, add ProfileFormFieldsScreenModel tests
