--- test/ui/profile/create_profile/widgets/create_profile_screen_model_test.mocks.dart
@@ -1,65 +1,0 @@
-// Mocks generated by Mockito 5.4.6 from annotations
-// in sport_matcher/test/ui/profile/create_profile/widgets/create_profile_screen_model_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i4;
-
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:sport_matcher/data/profile/domain/profile_domain.dart' as _i5;
-import 'package:sport_matcher/data/profile/repository/abstract_profiles_repository.dart'
-    as _i3;
-import 'package:sport_matcher/ui/core/utilities/validators/abstract_text_validator.dart'
-    as _i2;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: must_be_immutable
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-// ignore_for_file: invalid_use_of_internal_member
-
-/// A class which mocks [AbstractTextValidator].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockAbstractTextValidator extends _i1.Mock
-    implements _i2.AbstractTextValidator {
-  MockAbstractTextValidator() {
-    _i1.throwOnMissingStub(this);
-  }
-}
-
-/// A class which mocks [AbstractProfilesRepository].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockAbstractProfilesRepository extends _i1.Mock
-    implements _i3.AbstractProfilesRepository {
-  MockAbstractProfilesRepository() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i4.Future<void> addProfile(_i5.ProfileDomain? profile) =>
-      (super.noSuchMethod(
-            Invocation.method(#addProfile, [profile]),
-            returnValue: _i4.Future<void>.value(),
-            returnValueForMissingStub: _i4.Future<void>.value(),
-          )
-          as _i4.Future<void>);
-
-  @override
-  _i4.Future<_i5.ProfileDomain?> loadProfile() =>
-      (super.noSuchMethod(
-            Invocation.method(#loadProfile, []),
-            returnValue: _i4.Future<_i5.ProfileDomain?>.value(),
-          )
-          as _i4.Future<_i5.ProfileDomain?>);
-}
